PDF version of the presentation
This is a presentation that was used for Spring 2017 NAU Ugrads for Team B.A.M. from the computer science department. Ugrads is an even at NAU where capstone teams present their projects infront of an audience, anyone is allowed to attend. My capstone team was judge by an mechanical engineering professor for the presentation. Our over all Ugrads score was a 4.6 out of 5. This score was the poster score and this presentation score combined.