Poster Team B.A.M. used for Ugrads Presentation
On top of an oral presentation that we had to do for NAU Ugrads. We also had to have a poster and stand by it for two hours. We were judged by 5 judges. Our poster was very popular because of our project and we always had a lot of people around asking questions and ask to demo.